Resorted to using shorts for most simple geometry to minimize memory footprint. Improvements in the buffer API abstraction. Refactored opengl renderer to only use a single function per feature instead of using polymorphism. Vertex arrays now only hold onto one vertex buffer at a time, to keep it simple and avoid indexing issues with attributes. Greatly reduced vbo and ibo memory footprint by implementing a dynamic vbo and ibo resize (similar to how std::vector does it). Removed the erroneous data members of Object 2D, since normals and faces are not a thing unlike 3D.

+++ Wave/include/Renderer/gl_framebuffer.h
@@ -13,7 +13,7 @@ namespace Wave
   typedef struct Gl_framebuffer_viewport_data
   {
     std::shared_ptr<Vertex_array_buffer> vao;
-    uint32_t *ibo_data;
+    ushort *ibo_data;
     float *vbo_data;
   } Framebuffer_viewport_data;
